Completed integration with ESRI Geocoding API

+++ src/main/java/com/rutkowski/gisspatialdatamanager/types/geometries/PointGeometry.java
@@ -1,13 +1,11 @@
 package com.rutkowski.gisspatialdatamanager.types.geometries;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class PointGeometry {
  
-  private final BigDecimal latitude;
-  private final BigDecimal longitude;
+  private final double latitude;
+  private final double longitude;
 
 }
